requirements.txt
streamlit>=1.32
openai>=1.30
python-dotenv>=1.0
pandas>=2.0
matplotlib>=3.7
